moved more utils
