+++ eidoscommon/src/main/scala/org/clulab/wm/eidoscommon/utils/Timer.scala
@@ -1,4 +1,4 @@
-package org.clulab.wm.eidos.utils
+package org.clulab.wm.eidoscommon.utils
 
 import org.slf4j.{Logger, LoggerFactory}
