+++ src/main/scala/org/clulab/wm/eidos/apps/MakeRuleTSVs.scala
@@ -4,12 +4,11 @@ import org.clulab.odin.{Attachment, EventMention, Mention}
 import org.clulab.wm.eidos.EidosSystem
 import org.clulab.wm.eidos.attachments.{Decrease, Increase, Quantification}
 import org.clulab.wm.eidoscommon.utils.Closer.AutoCloser
-import org.clulab.wm.eidoscommon.utils.FileUtils
+import org.clulab.wm.eidoscommon.utils.{FileUtils, ThreadUtils}
 import org.clulab.wm.eidos.utils.FoundBy
 
 import scala.collection.Seq
 import scala.collection.mutable.ArrayBuffer
-import org.clulab.wm.eidos.utils.ThreadUtils
 import org.clulab.wm.eidoscommon.EidosParameters
