+++ src/main/scala/org/clulab/wm/eidos/apps/RoundTrip.scala
@@ -8,8 +8,7 @@ import org.clulab.wm.eidos.document.AnnotatedDocument
 import org.clulab.wm.eidos.document.AnnotatedDocument.Corpus
 import org.clulab.wm.eidos.serialization.jsonld.JLDCorpus
 import org.clulab.wm.eidos.serialization.jsonld.JLDDeserializer
-import org.clulab.wm.eidoscommon.utils.FileUtils
-import org.clulab.wm.eidos.utils.Timer
+import org.clulab.wm.eidoscommon.utils.{FileUtils, Timer}
 
 import scala.collection.Seq
