+++ src/main/scala/org/clulab/wm/eidos/apps/batch/ExtractCluMetaFromDirectoryFiltered.scala
@@ -14,8 +14,7 @@ import org.clulab.wm.eidos.EidosSystem
 import org.clulab.wm.eidos.groundings.grounders.EidosAdjectiveGrounder
 import org.clulab.wm.eidos.serialization.jsonld.JLDCorpus
 import org.clulab.wm.eidoscommon.utils.Closer.AutoCloser
-import org.clulab.wm.eidoscommon.utils.{FileEditor, FileUtils}
-import org.clulab.wm.eidos.utils.ThreadUtils
+import org.clulab.wm.eidoscommon.utils.{FileEditor, FileUtils, ThreadUtils}
 import org.clulab.wm.eidos.utils.meta.CluText
 
 object ExtractCluMetaFromDirectoryFiltered extends App {
