+++ src/test/scala/org/clulab/wm/eidos/system/TestDomainOntology.scala
@@ -5,8 +5,7 @@ import com.typesafe.config.{ConfigFactory, ConfigValueFactory}
 import org.clulab.wm.eidos.EidosSystem
 import org.clulab.wm.eidos.groundings._
 import org.clulab.wm.eidos.test.TestUtils._
-import org.clulab.wm.eidos.utils.Timer
-import org.clulab.wm.eidoscommon.utils.Canonicalizer
+import org.clulab.wm.eidoscommon.utils.{Canonicalizer, Timer}
 import org.clulab.wm.ontologies.CompactDomainOntology.CompactDomainOntologyBuilder
 import org.clulab.wm.ontologies.FastDomainOntology.FastDomainOntologyBuilder
 import org.clulab.wm.ontologies.{DomainOntology, FullTreeDomainOntology, HalfTreeDomainOntology}
